Funções principais finalizadas

+++ src/Vendas/Venda.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 public class Venda {
     protected ArrayList<String> dataVenda = new ArrayList<String>();
     protected ArrayList<String> nomeVeiculo = new ArrayList<String>();
-    protected ArrayList<Double> valorVeiculo = new ArrayList<Double>();
+    protected ArrayList<Float> valorVeiculo = new ArrayList<Float>();
     protected ArrayList<String> codigoCliente = new ArrayList<String>();
 
     public ArrayList<String> getDataVenda() {
@@ -18,10 +18,10 @@ public class Venda {
     public void setNomeVeiculo(String veiculo) {
         nomeVeiculo.add(veiculo);
     }
-    public ArrayList<Double> getValorVeiculo() {
+    public ArrayList<Float> getValorVeiculo() {
         return valorVeiculo;
     }
-    public void setValorVeiculo(Double valor) {
+    public void setValorVeiculo(Float valor) {
         valorVeiculo.add(valor);
     }
     public ArrayList<String> getCodigoCliente() {
